[add] cleaned up timeseries query runner
